==> 02_crosswalk.ipynb <==
# Crosswalk from onet Data to Swiss Jobs

onet soc to soc crosswalk: https://www.onetcenter.org/crosswalks.html  
soc to isco08 crosswalk: https://www.bls.gov/soc/ISCO_SOC_Crosswalk.xls  
ch isco 19: https://www.i14y.admin.ch/en/catalog/datasets/HCL_CH_ISCO_19_PROF_1_2/distributions  

## Content
1. [Crosswalk onet Osborne and Frey Imports](#crosswalk-import)
2. [Crosswalks](#crosswalks)
3. [safe to db](#safe-to-db)

# imports
import pandas as pd
from db import get_database

<a id='crosswalk-import'></a>
## Import Crosswalk files, data from Osborne & Frey and ONET data

# import onet crosswalk file
onet_file =  "files/crosswalk_ONETSOC_SOC.xlsx"
sheet =  "ONETSOC_SOC"

onetsoc_soc = pd.read_excel(io=onet_file, sheet_name=sheet , skiprows=3, dtype=str)

# rename cols
onetsoc_soc = onetsoc_soc.rename(columns={'onet-SOC 2019 Code': 'onetsoc19'})
onetsoc_soc = onetsoc_soc.rename(columns={'2018 SOC Code': '2010soc'})

# List of columns to drop
cols_to_drop = ['onet-SOC 2019 Title', '2018 SOC Title']

# Drop specified columns
onetsoc_soc = onetsoc_soc.drop(columns=cols_to_drop)
onetsoc_soc = onetsoc_soc.dropna()
print(onetsoc_soc.shape[0])

onetsoc_soc.head(2)

# import ISCO crosswalk file
file_name =  "files/crosswalk_ISCO_SOC.xls"
sheet =  "ISCO_SOC"

isco_soc = pd.read_excel(io=file_name, sheet_name=sheet , skiprows=6, dtype=str)

# rename cols
isco_soc = isco_soc.rename(columns={'ISCO-08 Code': 'isco08'})
isco_soc = isco_soc.rename(columns={'2010 SOC Code': '2010soc'})

# List of columns to drop
cols_to_drop = ['ISCO-08 Title EN', 'part', '2010 SOC Title', "Comment 8/17/11"]

# Drop specified columns
isco_soc = isco_soc.drop(columns=cols_to_drop)
isco_soc = isco_soc.dropna()
print(isco_soc.shape[0])

isco_soc.head(2)

# import CH crosswalk file
file_name =  "files/crosswalk_CH_ISCO_19.xlsx"
sheet =  "CH_ISCO_19"

ch_isco = pd.read_excel(io=file_name, sheet_name=sheet, skiprows=0, dtype=str)

# rename cols
ch_isco = ch_isco.rename(columns={'Berufsgattungen': 'isco08'})

# List of columns to drop
cols_to_drop = ["Berufshauptgruppen",'Berufsgruppen', 'Berufsuntergruppen', "Berufsarten", "Code"]

# Drop specified columns
ch_isco = ch_isco.drop(columns=cols_to_drop)
ch_isco = ch_isco.dropna()
print(ch_isco.shape[0])

ch_isco.head(2)

# import data from onet from mongodb
dbname = get_database()

collection_name = dbname["onet"]

data = collection_name.find()
onet = pd.DataFrame(list(data))
onet.head(2)

# import Source Frey / Osborne
fo =  pd.read_csv("files/FO.csv")

# Convert 'code' column to string
onet['code'] = onet['code'].astype(str)
fo["code"] = fo["code"].astype(str)

# Rename 'code' column to 'onetsoc19'
onet = onet.rename(columns={'code': 'onetsoc19'})
fo = fo.rename(columns={'code': '2010soc'})

print(onet.shape[0])

onet.head(2)

<a id='crosswalks'></a>
## Crosswalks

# crosswalk onet soc
joined = pd.merge(onet, onetsoc_soc, left_on='onetsoc19', right_on="onetsoc19", how='left')
joined = joined.dropna()

# print amount of cols
print(joined.shape[0])

# crosswalk soc
joined = pd.merge(joined, isco_soc, left_on='2010soc', right_on="2010soc", how='left')
fo_joined = pd.merge(fo, isco_soc, left_on='2010soc', right_on="2010soc", how='left')

# print amount of cols
print(joined.shape[0])

# crosswalk isco08
joined = pd.merge(joined, ch_isco, left_on='isco08', right_on="isco08", how='left')
fo_joined = pd.merge(fo_joined, ch_isco, left_on='isco08', right_on="isco08", how='left')

# print amount of cols
print(joined.shape[0])
print(fo_joined.shape[0])

# drop na values
joined = joined.dropna()
joined = joined.drop(columns=["_id"])
fo_joined = fo_joined.dropna()
print(joined.shape[0])

# drop further cols
joined = joined.drop(columns=["onetsoc19","2010soc"])
fo_joined = fo_joined.drop(columns=["2010soc","Occupation","Name_de","Rank"])
fo_joined = fo_joined.rename(columns={'Label': 'fo_computerisation',"Probability":"fo_probability"})

<a id="save-to-db"></a>
## safe to db

# save to mongodb
dbname = get_database()

collection_name = dbname["joined"]
fo_col = dbname["fo_joined"]

# Insert the dictionaries into the MongoDB collection
collection_name.insert_many(joined.to_dict("records"))
fo_col.insert_many(fo_joined.to_dict("records"))
