==> 03_convert.ipynb <==
# Convert Text to IDs

## Content
1. [Import data](#import)
2. [extract data and add IDs](#extract)
3. [save to csv](#save-to-csv)

# imports
import pandas as pd
from db import get_database

<a id='import'></a>
## Import data

# get data from mongodb
dbname = get_database()
collection = dbname["joined"]

documents = collection.find()
df = pd.DataFrame(list(documents))

df.head(2)

# create dfs
skills = pd.DataFrame()
abilities = pd.DataFrame()

<a id='extract'></a>
## Extract data and add IDs

# extract all skills from the onet data
skills = df['skills'].apply(lambda x: x['element']).explode().reset_index(drop=True)

# Convert the list of dictionaries into a DataFrame
skills_df = pd.json_normalize(skills)

# Drop duplicates based on 'id'
skills_df = skills_df.drop_duplicates(subset='id')

# Reset the index
skills_df = skills_df.reset_index(drop=True)

# drop cols
skills_df = skills_df.drop(columns=["related", "score.scale", "score.important", "score.value"])

# add numeric id col
skills_df['skill_id'] = range(1, len(skills_df) + 1)

skills_df.head(3)

# extract all abilities from the onet data
abilities = df['abilities'].apply(lambda x: x['element']).explode().reset_index(drop=True)

# Convert the list of dictionaries into a DataFrame
abilities_df = pd.json_normalize(abilities)

# Drop duplicates based on 'id'
abilities_df = abilities_df.drop_duplicates(subset='id')

# Reset the index
abilities_df = abilities_df.reset_index(drop=True)

# drop cols
abilities_df = abilities_df.drop(columns=["related", "score.scale", "score.important" ,"score.value"])

# add numeric id col
abilities_df['ability_id'] = range(1, len(abilities_df) + 1)

abilities_df.head(3)

<a id='save-to-csv'></a>
## Save to csv


abilities_df.to_csv("files/onet_abilities.csv", index=False)
skills_df.to_csv("files/onet_skills.csv", index=False)
