==> 05b_calculate_weights.ipynb <==
# Calculate Normalized Weights for Probability Calculation

## Content
1. [Import files](#import)
2. [Normalize Data](#normalize)
3. [Calculate Weights](#weights)
3. [Save Data](#save)

# imports
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from collections import Counter
from sklearn.metrics import * 
from sklearn.utils import resample
import matplotlib.colors as mcolors
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

<a id="import"></a>
## import

# import files
df = pd.read_csv("files/fo_smote.csv")

print(df.shape[0])
df.head(5)

<a id="normalize"></a>
## Normalize Data

# Initialize a scaler
scaler = StandardScaler()

# Define the columns to scale
cols_to_scale = [col for col in df.columns if col not in ['isco08', 'Name_de',"Berufshauptgruppe","fo_probability","fo_computerisation"]]

# Scale only these columns
df_scaled = df.copy()
df_scaled[cols_to_scale] = scaler.fit_transform(df[cols_to_scale])

df_scaled.head(5)

<a id="weights"></a>
## Calculate Coefficients

# Select columns that start with 's' or 'a' and the 'fo_probability' column
df_selected = df_scaled.filter(regex='^(s|a|fo_computerisation)')

# Define the dependent variable
y = df_selected['fo_computerisation']

# Define the independent variables
X = df_selected.drop('fo_computerisation', axis=1)

# Fit the model
model = LogisticRegression(C=0.3, penalty='l2', solver='newton-cg', random_state=42, max_iter=1000)
model.fit(X, y)

intercept = model.intercept_

# Get the regression coefficients
coefficients = pd.Series(model.coef_[0], index=X.columns)

# Create a dataframe from the coefficients outside the 95% and 99% confidence intervals with headers
df_coefficients = pd.DataFrame(list(coefficients.items()), columns=['id', 'coefficient'])

df_coefficients.head(2)

<a id="save"></a>
## Save the data

# Save the dataframes to csv files
df_coefficients.reset_index().rename(columns={'index': 'Variable', 0: 'Coefficient'}).sort_values('coefficient').drop(columns=["Variable"]).to_csv('files/coefficients.csv', index=False)
pd.DataFrame([intercept], columns=['intercept']).to_csv('files/intercept.csv', index=False)
