==> 08_comparison.ipynb <==
# Comparison to historical data

## Content
1. [Import Files](#import)
2. [Calculate Mean Importance per Fähigkeit per Berufshauptgruppe](#mean)
3. [Save File](#save)

# imports
import pandas as pd

<a id="import"></a>
## import

# import files
df_jobs = pd.read_csv('files/fo_swiss.csv')
not_automatable = [item[0] for item in pd.read_csv("files/not_automatable.csv").values.tolist()]

<a id="mean"></a>
## Calculate Mean of Importance per Fähigkeit per Berufshauptgrubbe

# Select the columns in df_jobs that are also in not_automatable
selected_columns = [col for col in df_jobs.columns if col in not_automatable]

# Group by 'Berufshauptgruppe' and calculate the mean value for the selected columns
mean_values_per_berufshauptgruppe = df_jobs.groupby('Berufshauptgruppe')[selected_columns].mean()

<a id="save"></a>
## Save File

# Save the DataFrame as a CSV file
mean_values_per_berufshauptgruppe.to_csv('files/berufshauptgruppe_mean_importance.csv')
